+++ WordPressKitTests/WordPressComRestApiTests.swift
@@ -81,11 +81,11 @@ class WordPressComRestApiTests: XCTestCase {
 
     func testBaseUrl() {
         let defaultApi = WordPressComRestApi()
-        XCTAssertEqual(defaultApi.baseURLString, "https://public-api.wordpress.com/")
+        XCTAssertEqual(defaultApi.baseURL.absoluteString, "https://public-api.wordpress.com/")
         XCTAssertTrue(defaultApi.buildRequestURLFor(path: "/path")!.hasPrefix("https://public-api.wordpress.com/path"))
 
-        let localhostApi = WordPressComRestApi(baseUrlString: "http://localhost:8080")
-        XCTAssertEqual(localhostApi.baseURLString, "http://localhost:8080")
+        let localhostApi = WordPressComRestApi(baseURL: URL(string: "http://localhost:8080")!)
+        XCTAssertEqual(localhostApi.baseURL.absoluteString, "http://localhost:8080")
         XCTAssertTrue(localhostApi.buildRequestURLFor(path: "/path")!.hasPrefix("http://localhost:8080/path"))
     }